convert to tensors again

+++ main.py
@@ -45,10 +45,10 @@ print(X)
 # Test train split for pytorch: 80 train, 20 test
 X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=41)
 
-# Convert X features to float tensors (?)
+# Convert X features to float tensors (?) integers
 X_train = torch.FloatTensor(X_train)
 X_test = torch.FloatTensor(X_test)
 
-# Convert y labels to float tensors (?)
-y_train = torch.FloatTensor(y_train)
-y_test = torch.FloatTensor(y_test)
+# Convert y labels to long tensors (?) 64 bit integers
+y_train = torch.LongTensor(y_train)
+y_test = torch.LongTensor(y_test)
